+++ tests/unit/PaymentMethods/CreditCardPaymentMethodTest.php
@@ -9,6 +9,7 @@ use Heidelpay\PhpPaymentApi\PaymentMethods\CreditCardPaymentMethod;
 use Heidelpay\Tests\PhpPaymentApi\Helper\BasePaymentMethodTest;
 use Heidelpay\PhpPaymentApi\Constants\TransactionType;
 use Heidelpay\PhpPaymentApi\Constants\TransactionMode;
+use Heidelpay\PhpPaymentApi\Constants\FrontendMode;
 
 /**
  * This test class verifies the special functionality of the CreditCardPaymentMethod not covered in
@@ -216,7 +217,7 @@ class CreditCardPaymentMethodTest extends BasePaymentMethodTest
                 'CRITERION.SDK_VERSION' => ApiConfig::SDK_VERSION,
                 'FRONTEND.CSS_PATH' => self::CSS_PATH,
                 'FRONTEND.ENABLED' => 'FALSE',
-                'FRONTEND.MODE' => 'WHITELABEL',
+                'FRONTEND.MODE' => FrontendMode::FRONTEND_MODE_WHITELABEL,
                 'FRONTEND.PAYMENT_FRAME_ORIGIN' => self::PAYMENT_FRAME_ORIGIN,
                 'FRONTEND.PREVENT_ASYNC_REDIRECT' => $preventAsyncRedirect,
                 'IDENTIFICATION.SHOPPERID' => $shopperId,
@@ -298,7 +299,7 @@ class CreditCardPaymentMethodTest extends BasePaymentMethodTest
                 'CRITERION.SDK_VERSION' => ApiConfig::SDK_VERSION,
                 'FRONTEND.CSS_PATH' => self::CSS_PATH,
                 'FRONTEND.ENABLED' => 'FALSE',
-                'FRONTEND.MODE' => 'WHITELABEL',
+                'FRONTEND.MODE' => FrontendMode::FRONTEND_MODE_WHITELABEL,
                 'FRONTEND.PAYMENT_FRAME_ORIGIN' => self::PAYMENT_FRAME_ORIGIN,
                 'FRONTEND.PREVENT_ASYNC_REDIRECT' => $preventAsyncRedirect,
                 'IDENTIFICATION.SHOPPERID' => $shopperId,
@@ -374,7 +375,7 @@ class CreditCardPaymentMethodTest extends BasePaymentMethodTest
             'CRITERION.SDK_VERSION' => ApiConfig::SDK_VERSION,
             'FRONTEND.CSS_PATH' => self::CSS_PATH,
             'FRONTEND.ENABLED' => $frontendEnabled,
-            'FRONTEND.MODE' => 'WHITELABEL',
+            'FRONTEND.MODE' => FrontendMode::FRONTEND_MODE_WHITELABEL,
             'FRONTEND.PAYMENT_FRAME_ORIGIN' => self::PAYMENT_FRAME_ORIGIN,
             'FRONTEND.PREVENT_ASYNC_REDIRECT' => $preventAsyncRedirect,
             'IDENTIFICATION.SHOPPERID' => $shopperId,
@@ -473,7 +474,7 @@ class CreditCardPaymentMethodTest extends BasePaymentMethodTest
             'CRITERION.SDK_VERSION' => ApiConfig::SDK_VERSION,
             'FRONTEND.CSS_PATH' => self::CSS_PATH,
             'FRONTEND.ENABLED' => $frontendEnabled,
-            'FRONTEND.MODE' => 'WHITELABEL',
+            'FRONTEND.MODE' => FrontendMode::FRONTEND_MODE_WHITELABEL,
             'FRONTEND.PAYMENT_FRAME_ORIGIN' => self::PAYMENT_FRAME_ORIGIN,
             'FRONTEND.PREVENT_ASYNC_REDIRECT' => $preventAsyncRedirect,
             'IDENTIFICATION.SHOPPERID' => $shopperId,