+++ CHANGELOG.md
@@ -3,10 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## X.X.X
-
-## Changed
+## [v1.5.1][v1.5.1]
+### Changed
+- Replaced warning suppression in mapFromPost method.
 - Visibility of *Parent methods in `CreditCardPaymentMethod` and `DebitCardPaymentMethods`
+- Prohibit setting the mode parameter of the frontend group.
+
+### Fixed
+- Parameter names in riskinformation group.
 
 ## [v1.5.0][v1.5.0]
 ### Added
@@ -115,3 +119,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [v1.4.0]: https://github.com/heidelpay/php-payment-api/compare/v1.3.0...v1.4.0
 [v1.4.1]: https://github.com/heidelpay/php-payment-api/compare/v1.4.0...v1.4.1
 [v1.5.0]: https://github.com/heidelpay/php-payment-api/compare/v1.4.0...v1.5.0
+[v1.5.1]: https://github.com/heidelpay/php-payment-api/compare/v1.5.0...v1.5.1